Combine Providers and remove extra index.php/config. Add back in error handlers.
