--- src/TransactionService/config/example.settings.yml
@@ -1,13 +1,0 @@
-# Rename this file to settings.yml for production use or
-# rename it to settings.dev.yml to be used with $app['debug'] == TRUE
-islandora:
-  apiVersion: 0.0.1
-  fedoraProtocol: http
-  fedoraHost: "localhost:8080"
-  fedoraPath: /fcrepo/rest
-  tripleProtocol: http
-  tripleHost: "localhost:8080"
-  triplePath: /bigdata/sparql
-  resourceIdRegex: "(?:[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12})?"
-  # This domain is used as namespace (hashed) when generating UUID V5 identifiers, replace with your own
-  defaultNamespaceDomainUuuidV5: "www.islandora.ca"
